Feat/api

+++ backend/SharpTask.Application/Services/NoteCommandService.cs
@@ -8,16 +8,21 @@ using SharpTask.Domain.Interfaces;
 
 namespace SharpTask.Application.Services;
 
+/// <summary>
+/// Servicio de comandos para las notas,
+/// encargado de manejar la lógica de negocio relacionada con la creación,
+/// actualización y eliminación de notas asociadas a tareas.
+/// </summary>
 public class NoteCommandService : INoteCommandService
 {
     private readonly IDateTimeProvider _dateTimeProvider;
     private readonly INoteRepository _noteRepo;
     private readonly ITaskRepository _taskRepo;
 
-    /// <summary>
+    /// <remarks>
     /// Constructor del servicio de comandos de notas que recibe una instancia del repositorio de notas
     /// para acceder a los datos de las notas y proporcionar la funcionalidad de comando relacionada con las notas.
-    /// </summary>
+    /// </remarks>
     /// <param name="noteRepo">La instancia del repositorio de notas.</param>
     /// <param name="taskRepo">La instancia del repositorio de tareas.</param>
     /// <param name="dateTimeProvider">La instancia del proveedor de fechas y horas.</param>
@@ -32,10 +37,10 @@ public class NoteCommandService : INoteCommandService
         _dateTimeProvider = dateTimeProvider;
     }
 
-    /// <summary>
+    /// <remarks>
     /// Crea una nueva nota asociada a una tarea específica utilizando los datos proporcionados en el DTO de solicitud.
     /// Mapea la nota creada a un DTO de respuesta para ser consumido por el frontend y lo devuelve.
-    /// </summary>
+    /// </remarks>
     /// <param name="taskId">El ID de la tarea a la que se asociará la nota.</param>
     /// <param name="request">El DTO de solicitud con los datos de la nota a crear.</param>
     /// <returns>El DTO de respuesta con los datos de la nota creada.</returns>
@@ -67,22 +72,29 @@ public class NoteCommandService : INoteCommandService
         return createdNote.Adapt<NoteResponseDto>();
     }
 
-    /// <summary>
+    /// <remarks>
     /// Actualiza una nota existente utilizando los datos proporcionados en el DTO de solicitud.
     /// Verifica si la nota existe antes de intentar actualizarla y devuelve null si no se puede actualizar.
     /// Mapea la nota actualizada a un DTO de respuesta para ser consumido por el frontend y lo devuelve.
-    /// </summary>
+    /// </remarks>
+    /// <param name="taskId">El ID de la tarea a la que pertenece la nota.</param>
     /// <param name="id">El ID de la nota a actualizar.</param>
     /// <param name="request">El DTO de solicitud con los datos de la nota a actualizar.</param>
     /// <returns>El DTO de respuesta con los datos de la nota actualizada o null si no se puede actualizar.</returns>
-    public async Task<NoteResponseDto?> UpdateNoteAsync(Guid id, UpdateNoteRequestDto request)
+    public async Task<NoteResponseDto?> UpdateNoteAsync(
+        Guid taskId,
+        Guid id,
+        UpdateNoteRequestDto request
+    )
     {
         // Obtenemos la nota existente por su ID para verificar si existe antes de intentar actualizarla
         var existingNote = await _noteRepo.GetByIdAsync(id);
 
-        // Si la nota no existe, devolvemos null para indicar que no se pudo actualizar
-        if (existingNote == null)
+        // Verificamos que la nota exista y que esté asociada a la tarea especificada antes de intentar actualizarla
+        if (existingNote is null || existingNote.TaskId != taskId)
+        {
             return null;
+        }
 
         // Mapeamos los datos del DTO de solicitud a la nota existente
         // para actualizar sus propiedades con los nuevos valores proporcionados
@@ -98,13 +110,24 @@ public class NoteCommandService : INoteCommandService
         return updatedNote?.Adapt<NoteResponseDto>();
     }
 
-    /// <summary>
+    /// <remarks>
     /// Elimina una nota por su ID utilizando el repositorio de notas.
-    /// </summary>
+    /// </remarks>
+    /// <param name="taskId">El ID de la tarea a la que pertenece la nota.</param>
     /// <param name="id">El ID de la nota a eliminar.</param>
     /// <returns>True si la nota existe y se elimino correctamente, false en caso contrario.</returns>
-    public async Task<bool> DeleteNoteAsync(Guid id)
+    public async Task<bool> DeleteNoteAsync(Guid taskId, Guid id)
     {
+        // Obtenemos la nota existente por su ID para verificar si existe antes de intentar eliminarla
+        var existingNote = await _noteRepo.GetByIdAsync(id);
+
+        // Verificamos que la nota exista y que esté asociada a la tarea especificada antes de intentar eliminarla
+        if (existingNote is null || existingNote.TaskId != taskId)
+        {
+            return false;
+        }
+
+        // Eliminamos la nota utilizando el repositorio de notas y devolvemos el resultado de la operación
         return await _noteRepo.DeleteAsync(id);
     }
 }
